Hand_written_digit_analyser/requirements.txt
flask>=3.0.0
tensorflow>=2.13.0
numpy
matplotlib
scikit-learn
seaborn
pillow
